Change stats command syntax; use config for local queries
The local configuration will be used to configure client options when querying
the locally running nvremoted server.
For this to work, TLS certificate verification needs to be disabled,
since the hostname in the certificate will most likely not be "127.0.0.1."

The stats password can now additionally be set in the NVREMOTED_STATS_PASSWORD environment variable. The password in the local config will no longer be used when connecting to remote servers.

Add a --server-certificate option, to trust a server's self-signed certificate.
When this option is used, the host CA store will not be used.